Changed setPrecio to double

+++ src/main/java/almacen/model/Producto.java
@@ -46,7 +46,7 @@ public class Producto {
         return precio;
     }
 
-    public void setPrecio(int precio) {
+    public void setPrecio(double precio) {
         this.precio = precio;
     }
